backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
faiss-cpu
sentence-transformers
requests
beautifulsoup4
pypdf
youtube-transcript-api
python-dotenv
